+++ backend/src/neuroagent/app/config.py
@@ -1,11 +1,16 @@
 """Configuration."""
 
+import json
+import logging
 import os
+from pathlib import Path
 from typing import Any, Literal, Optional
 
 from dotenv import dotenv_values
-from pydantic import BaseModel, ConfigDict, Field, SecretStr, model_validator
+from pydantic import BaseModel, ConfigDict, SecretStr, model_validator
 from pydantic_settings import BaseSettings, SettingsConfigDict
+
+logger = logging.getLogger(__name__)
 
 
 class SettingsAgent(BaseModel):
@@ -26,8 +31,6 @@ class SettingsStorage(BaseModel):
     access_key: SecretStr | None = None
     secret_key: SecretStr | None = None
     expires_in: int = 600
-    brain_region_hierarchy_key: str = "shared/brainregion_hierarchy.json"
-    cell_type_hierarchy_key: str = "shared/celltypes_hierarchy.json"
 
     model_config = ConfigDict(frozen=True)
 
@@ -83,30 +86,6 @@ class SettingsLiterature(BaseModel):
     model_config = ConfigDict(frozen=True)
 
 
-class SettingsTrace(BaseModel):
-    """Trace tool settings."""
-
-    search_size: int = 10
-
-    model_config = ConfigDict(frozen=True)
-
-
-class SettingsKGMorpho(BaseModel):
-    """KG Morpho settings."""
-
-    search_size: int = 3
-
-    model_config = ConfigDict(frozen=True)
-
-
-class SettingsGetMEModel(BaseModel):
-    """Get ME Model settings."""
-
-    search_size: int = 10
-
-    model_config = ConfigDict(frozen=True)
-
-
 class SettingsBlueNaaS(BaseModel):
     """BlueNaaS settings."""
 
@@ -121,33 +100,6 @@ class SettingsEntityCore(BaseModel):
     model_config = ConfigDict(frozen=True)
 
 
-class SettingsKnowledgeGraph(BaseModel):
-    """Knowledge graph API settings."""
-
-    base_url: str = "https://www.openbraininstitute.org/api/nexus/v1"
-    model_config = ConfigDict(frozen=True)
-
-    @property
-    def url(self) -> str:
-        """Knowledge graph search url."""
-        return f"{self.base_url}/search/query/"
-
-    @property
-    def sparql_url(self) -> str:
-        """Knowledge graph view for sparql query."""
-        return f"{self.base_url}/views/neurosciencegraph/datamodels/https%3A%2F%2Fbluebrain.github.io%2Fnexus%2Fvocabulary%2FdefaultSparqlIndex/sparql"
-
-    @property
-    def class_view_url(self) -> str:
-        """Knowledge graph view for ES class query."""
-        return f"{self.base_url}/views/neurosciencegraph/datamodels/https%3A%2F%2Fbbp.epfl.ch%2Fneurosciencegraph%2Fdata%2Fviews%2Fes%2Fdataset/_search"
-
-    @property
-    def hierarchy_url(self) -> str:
-        """Knowledge graph url for brainregion/celltype files."""
-        return "http://bbp.epfl.ch/neurosciencegraph/ontologies/core"
-
-
 class SettingsTools(BaseModel):
     """Database settings."""
 
@@ -155,9 +107,6 @@ class SettingsTools(BaseModel):
     obi_one: SettingsObiOne = SettingsObiOne()
     bluenaas: SettingsBlueNaaS = SettingsBlueNaaS()
     entitycore: SettingsEntityCore = SettingsEntityCore()
-    trace: SettingsTrace = SettingsTrace()
-    kg_morpho_features: SettingsKGMorpho = SettingsKGMorpho()
-    me_model: SettingsGetMEModel = SettingsGetMEModel()
     web_search: SettingsWebSearch = SettingsWebSearch()
 
     model_config = ConfigDict(frozen=True)
@@ -250,16 +199,13 @@ class MCPServerConfig(BaseModel):
 class SettingsMCP(BaseModel):
     """Settings for the MCP."""
 
-    servers: dict[str, MCPServerConfig] = Field(
-        default_factory=dict,
-    )
+    servers: dict[str, MCPServerConfig] | None = None
 
 
 class Settings(BaseSettings):
     """All settings."""
 
     tools: SettingsTools = SettingsTools()
-    knowledge_graph: SettingsKnowledgeGraph = SettingsKnowledgeGraph()
     agent: SettingsAgent = SettingsAgent()  # has no required
     db: SettingsDB = SettingsDB()  # has no required
     openai: SettingsOpenAI = SettingsOpenAI()  # has no required
@@ -278,6 +224,38 @@ class Settings(BaseSettings):
         frozen=True,
         extra="ignore",
     )
+
+    @model_validator(mode="before")
+    @classmethod
+    def parse_mcp_servers(cls, data: Any) -> Any:
+        """Read the `mcp.json` file and parse mcp servers."""
+        # Read the server config
+        with (Path(__file__).parent.parent / "mcp.json").open() as f:
+            servers = f.read()
+
+        # Replace placeholders with secret values in server config
+        if "mcp" in data.keys():
+            for secret_key, secret_value in data["mcp"].get("secrets", {}).items():
+                placeholder = f"NEUROAGENT_MCP__SECRETS__{secret_key.upper()}"
+                replacement = secret_value or ""
+                servers = servers.replace(placeholder, replacement)
+
+        mcps: dict[str, dict[str, Any]] = {"servers": {}}
+        # Parse and set the mcp servers
+        for server, config in json.loads(servers).items():
+            # If a secret is not set, do not include the associated server
+            if config.get("env") and any(
+                "NEUROAGENT_MCP__SECRETS__" in value for value in config["env"].values()
+            ):
+                logger.warning(
+                    f"MCP server {server} deactivated because some of its secrets were not provided."
+                )
+                continue
+            mcps["servers"][server] = config
+
+        data["mcp"] = mcps
+
+        return data
 
 
 # Load the remaining variables into the environment